Multi-Tello-Formation/network_setup.py
import yaml
import socket
import netifaces
import time
from typing import List, Dict

class NetworkConfigurator:
    def __init__(self, config_file: str = 'network_config.yaml'):
        self.config_file = config_file
        self.config = self._load_config()
        
    def _load_config(self) -> Dict:
        """Load configuration from YAML file"""
        with open(self.config_file, 'r') as f:
            return yaml.safe_load(f)
    
    def get_drone_configs(self) -> List[Dict]:
        """Get list of drone configurations"""
        return self.config['drones']
    
    def get_network_settings(self) -> Dict:
        """Get network settings"""
        return self.config['network']
    
    def scan_for_drones(self) -> List[str]:
        """Scan for available Tello drones in the network"""
        available_networks = []
        # Note: This function needs to be implemented based on your system's
        # wireless network scanning capabilities
        print("Please implement network scanning based on your system's capabilities")
        return available_networks
    
    def connect_to_drone(self, ssid: str, password: str = None) -> bool:
        """Connect to a specific drone's WiFi network"""
        # Note: This function needs to be implemented based on your system's
        # network connection capabilities
        print(f"Attempting to connect to drone network: {ssid}")
        print("Please implement network connection based on your system's capabilities")
        return False

def main():
    # Initialize network configurator
    config = NetworkConfigurator()
    
    # Get drone configurations
    drones = config.get_drone_configs()
    network_settings = config.get_network_settings()
    
    print("Loaded configuration:")
    print(f"Number of drones: {len(drones)}")
    print(f"Network settings: {network_settings}")
    
    # Scan for available drones
    available_networks = config.scan_for_drones()
    print(f"Available drone networks: {available_networks}")
    
    # Example: Connect to each drone
    for drone in drones:
        if drone['ssid']:
            success = config.connect_to_drone(drone['ssid'], drone['password'])
            if success:
                print(f"Successfully connected to {drone['name']}")
            else:
                print(f"Failed to connect to {drone['name']}")

if __name__ == "__main__":
    main()
