Multi-Tello-Formation/requirements.txt
pyyaml>=6.0
netifaces>=0.11.0
djitellopy>=2.5.0
